requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.1.0
blake3==1.0.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
compressed-tensors==0.9.1
contourpy==1.3.0
cupy-cuda12x==13.4.0
cycler==0.12.1
datasets==3.3.2
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.16
email-validator==2.2.0
eval-type-backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
gguf==0.10.0
gitdb==4.0.12
gitpython==3.1.44
h11==0.14.0
h5py==3.13.0
hf-transfer==0.1.9
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
importlib-metadata==8.6.1
importlib-resources==6.5.2
iniconfig==2.0.0
interegular==0.3.3
isort==5.12.0
jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lark==1.2.2
lazy-loader==0.4
llvmlite==0.43.0
lm-format-enforcer==0.10.11
markupsafe==2.1.5
matplotlib==3.9.4
mistral-common==1.5.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.65.2
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines-core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
pillow==11.0.0
pluggy==1.5.0
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.2
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.6
pydantic-core==2.27.2
pydocstyle==6.3.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.2.1
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich-toolkit==0.13.2
rpds-py==0.23.1
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
starlette==0.46.0
sympy==1.13.1
tifffile==2024.8.30
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0
triton==3.1.0
typeguard==4.4.2
typer==0.15.2
typing-extensions==4.12.2
tyro==0.9.16
tzdata==2025.1
unsloth==2024.11.10
unsloth-zoo==2024.11.5
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.7.3
watchfiles==1.0.4
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
